Fix circular dependency in unit tests
