+++ src/shared/lib/GenericAssayUtils/GenericAssayCommonUtils.spec.ts
@@ -5,13 +5,13 @@ import {
     getGenericAssayMetaPropertyOrDefault,
     COMMON_GENERIC_ASSAY_PROPERTY,
     deriveDisplayTextFromGenericAssayType,
-    GenericAssayTypeConstants,
     formatGenericAssayCompactLabelByNameAndId,
     filterGenericAssayEntitiesByGenes,
     makeGenericAssayPlotsTabOption,
 } from './GenericAssayCommonUtils';
 import { getServerConfig } from 'config/config';
 import ServerConfigDefaults from 'config/serverConfigDefaults';
+import { GenericAssayTypeConstants } from 'shared/lib/GenericAssayUtils/GenericAssayConfig';
 
 describe('GenericAssayCommonUtils', () => {
     describe('makeGenericAssayOption()', () => {
@@ -373,8 +373,8 @@ describe('GenericAssayCommonUtils', () => {
                 filterGenericAssayEntitiesByGenes(
                     [genericAssayEntity],
                     TARGET_GENE_LIST
-                ),
-                []
+                ).length,
+                0
             );
         });
     });
